convert img2img+alpha into inpainting
