+++ tests/test_horde_inference_painting.py
@@ -46,7 +46,29 @@ class TestHordeInference:
             "n_iter": 1,
             "model": "stable_diffusion_inpainting",
             "source_image": Image.open("images/test_inpaint_alpha.png"),
-            "source_processing": "inpainting",
+            "source_processing": "img2img",
+        }
+        assert self.horde is not None
+        pil_image = self.horde.basic_inference(data)
+        assert pil_image is not None
+        pil_image.save("images/horde_inpainting_alpha_mask.webp", quality=90)
+
+    def test_img2img_alpha_to_inpainting_convert(self):
+        data = {
+            "sampler_name": "k_dpmpp_2m",
+            "cfg_scale": 7.5,
+            "denoising_strength": 1,
+            "seed": 3,
+            "height": 512,
+            "width": 512,
+            "karras": True,
+            "clip_skip": 1,
+            "prompt": "a mecha robot sitting on a bench",
+            "ddim_steps": 20,
+            "n_iter": 1,
+            "model": "Deliberate",
+            "source_image": Image.open("images/test_inpaint_alpha.png"),
+            "source_processing": "img2img",
         }
         assert self.horde is not None
         pil_image = self.horde.basic_inference(data)
